+++ index.html
@@ -230,5 +230,7 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
+  <script type="module" src="js/main.js"></script>
+
 </body>
 </html>
