+++ js/utils/get-random-positive-integer.js
@@ -27,4 +27,4 @@ function getRandomPositiveInteger (a, b) {
   return Math.floor(result);
 }
 
-getRandomPositiveInteger ();
+export {getRandomPositiveInteger};
