+++ package.json
@@ -1,4 +1,5 @@
 {
+  "type": "module",
   "name": "kekstagram",
   "version": "23.0.0",
   "private": true,
